% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoMix_two_component_mixture_models.R
\docType{data}
\name{dfTumorIsof}
\alias{dfTumorIsof}
\title{Human Breast Cancer RNA-sequencing data from TCGA - Tumor Tissue}
\description{
Human Breast Cancer RNA-sequencing data from TCGA - Tumor Tissue
}
\references{
\url{https://gdc.cancer.gov/}
}
\author{
Daniel Pique \email{daniel.pique@med.einstein.yu.edu}
}
\keyword{Breast}
\keyword{RNA}
\keyword{Tumor}
\keyword{expression,}
