% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoMix_two_component_mixture_models.R
\docType{data}
\name{exprTumIsof}
\alias{exprTumIsof}
\title{Human Breast Cancer RNA-sequencing data from TCGA - Tumor Tissue}
\description{
These RNA-sequencing expression data were obtained from the Cancer Genome
Atlas project (now the Genomic Data Commons (GDC)). The sequencing data were
generated from breast carcinoma samples from 113 patients. Quantification of
RNA expression values was performed using standard GDC pipelines. The
expression values are reported in transcripts per million reads. Out of an
initial 73,599 RNA transcripts, 700 are included as part of this dataset.
These 700 transcripts represent a random subset of the transcripts with at
least 20% non-zero expression values across all patient samples. Rows contain
anonymized patient identifiers, while columns contain UCSC gene symbols.
}
\references{
\url{https://gdc.cancer.gov/}
}
\author{
Daniel Pique \email{daniel.pique@med.einstein.yu.edu}
}
\keyword{Breast}
\keyword{RNA}
\keyword{Tumor}
\keyword{expression,}
