% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoMix_examples.R
\name{oncoMixBimodal}
\alias{oncoMixBimodal}
\title{Creating an schematic of a 2-component mixture model}
\usage{
oncoMixBimodal(oe_means = c(3, 7))
}
\arguments{
\item{means}{Set the values for the difference between parameter means}
}
\description{
This function allows you to generate a plot
}
\examples{
oncoMixBimodal(oe_means = c(3,7))
oncoMixBimodal(oe_means = c(3,10))
}
\keyword{idealized,}
\keyword{oncoMix,}
\keyword{theoretical}
