% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoMix_examples.R
\name{oncoMixIdeal}
\alias{oncoMixIdeal}
\title{Creating an ideal oncomix gene}
\usage{
oncoMixIdeal(means = c(3, 7))
}
\arguments{
\item{means}{Set the difference between parameter means for the
overexpressed (oe) group. Defaults to c(3,7)}
}
\value{
Returns a ggplot object that shows the statistical model for an
idealized/theoretical oncogene candidate mRNA that is overexpressed in a
subset of tumors
}
\description{
This function allows you to generate a plot
}
\examples{
oncoMixIdeal(means=c(3,10))
oncoMixIdeal(means=c(2,18.5))
}
\keyword{idealized,}
\keyword{oncoMix,}
\keyword{theoretical}
