% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoMix_two_component_mixture_models.R
\name{plotGeneHist}
\alias{plotGeneHist}
\title{Plot a histogram of gene expression values from tumor and adjacent normal tissue.}
\usage{
plotGeneHist(mmParams, dfNml, dfTumor, isof)
}
\arguments{
\item{mmParams}{The output from the getMixModelParams function.}

\item{dfNml}{The normal dataframe.}

\item{dfTumor}{The tumor dataframe.}

\item{isof}{The gene isoform to visualize}
}
\value{
Returns a histogram of the gene expression values from the two groups.
}
\description{
This function allows you to plot a histogram of gene expression values
from tumor and adjacent normal tissue with the option of including the
best fitting Gaussian curve.
}
\examples{
data(dfTumorIsof, dfNmlIsof, package="oncomix")
mmParams = mixModelParams(dfNmlIsof, dfTumorIsof)
isof = rownames(mmParams)[1]
plotGeneHist(mmParams, dfNmlIsof, dfTumorIsof, isof)
}
\seealso{
\code{\link{mixModelParams}}
}
\keyword{Gaussian,}
\keyword{oncoMix,}
\keyword{two-component}
\keyword{visualization,}
