% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoMix_two_component_mixture_models.R
\name{plotSelectivityIndex}
\alias{plotSelectivityIndex}
\title{Plot the selectivity Index from getMixModelParams}
\usage{
plotSelectivityIndex(mmParams, dfNml)
}
\arguments{
\item{mmParams}{The output from the mixModelParams function. Will utilize the deltaMu2
and deltaMu1 rows}

\item{dfNml}{The normal dataframe. Will be used to calculate the SI.}
}
\value{
Returns a vector, equal in length to the number of genes. 1 SI per gene.
}
\description{
This function allows you to generate the parameters for two 2-component mixture models
with equal variances
}
\examples{
selectivityIndex(mmParams, dfNml)
}
\seealso{
\code{\link{mixModelParams}}
}
\keyword{oncoMix,}
\keyword{two-component}
\keyword{visualization,}
