% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoMix_two_component_mixture_models.R
\docType{data}
\name{queryRes}
\alias{queryRes}
\title{Oncogene Database Mapping Gene Symbol to UCSC ID (kgID)}
\description{
Oncogene Database Mapping Gene Symbol to UCSC ID (kgID)
}
\references{
\url{http://ongene.bioinfo-minzhao.org/ongene_human.txt}
}
\author{
Daniel Pique \email{daniel.pique@med.einstein.yu.edu}
}
\keyword{Database}
\keyword{Oncogenes,}
