% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoMix_two_component_mixture_models.R
\name{toMatrix}
\alias{toMatrix}
\title{Convert SummarizedExperiment or Dataframe to Matrix}
\usage{
toMatrix(m)
}
\arguments{
\item{m}{Can be a matrix, a data.frame, a DataFrame, or
SummarizedExperiment object.}
}
\value{
A matrix of expression values
}
\description{
This internal function converts SummarizedExperiment objects and dataframes
(both S3 and S4) to matrices of expression values. Used within oncomix
functions to convert all matrix-like objects to the matrix class.
}
\examples{
m <- as.data.frame(matrix(data=rgamma(n=150, shape=2, rate=2),
nrow=10, ncol=15))
m <- toMatrix(m)
}
\keyword{internal}
