% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoMix_two_component_mixture_models.R
\name{topGeneTable}
\alias{topGeneTable}
\title{Identify the top N genes based on a score}
\usage{
topGeneTable(mmParams, N = nrow(mmParams))
}
\arguments{
\item{mmParams}{The output from the mixModelParams function.}

\item{N}{An integer specified by the user to indicate how many genes they would like to return.}
}
\value{
Returns a dataframe containing the top N genes (where N is specified by the user)
computed using a custom score that incorporates differences between component means and the variance.
}
\description{
This function allows you to identify the top N genes that most
closely match the distributional profiles of a theoretical oncogene.
}
\examples{
topGeneQuants(mmParams, N = 10)
}
\seealso{
\code{\link{mixModelParams}}
}
\keyword{subsetting}
