% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoMix_examples.R
\name{oncoMixIdeal}
\alias{oncoMixIdeal}
\title{Creating an ideal oncoMix gene}
\usage{
oncoMixIdeal(oe_means = c(3, 7))
}
\arguments{
\item{oe_means}{Set the difference between parameter means for the overexpressed (oe) group. Defaults to c(3,7)}
}
\description{
This function allows you to generate a plot
}
\examples{
oncoMixIdeal(oe_means = c(3,10))
oncoMixIdeal(oe_means = c(2,18.5))
}
\keyword{idealized,}
\keyword{oncoMix,}
\keyword{theoretical}
