% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoMix_examples.R
\name{oncoMixTraditionalDE}
\alias{oncoMixTraditionalDE}
\title{Creating an schematic of a traditional differential expression experiment}
\usage{
oncoMixTraditionalDE(means = c(3, 7))
}
\arguments{
\item{means}{Set the values for the difference between parameter means}
}
\description{
This function allows you to generate a schematic of the assumptions of a traditional DE
expermiment between two known groups.
}
\examples{
oncoMixBimodal(means = c(3,7))
oncoMixBimodal(means = c(3,10))
}
\keyword{differential}
\keyword{expression}
\keyword{idealized,}
\keyword{oncoMix,}
\keyword{theoretical,}

