% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoMix_two_component_mixture_models.R
\docType{data}
\name{queryRes}
\alias{queryRes}
\title{Oncogene Database Mapping Gene Symbol to UCSC ID (kgID)}
\description{
These data were downloaded in September 2017 from the url listed below and
represent a mapping between gene symbols in ongene, an oncogene database
curated from the scientific literature, and gene identifiers from the
University of California, Santa Cruz's genomic database.
}
\references{
\url{http://ongene.bioinfo-minzhao.org/ongene_human.txt}
}
\author{
Min Zhao \email{mzhao@usc.edu.au}
}
\keyword{Database}
\keyword{Oncogenes,}
