% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoMix_two_component_mixture_models.R
\name{topGeneQuants}
\alias{topGeneQuants}
\title{Identify genes that meet pre-specified quantiles}
\usage{
topGeneQuants(mmParams, deltMu2Thresh = 90, deltMu1Thresh = 10,
  siThresh = 0.99)
}
\arguments{
\item{mmParams}{The output from the mixModelParams function.}

\item{deltMu2Thresh}{The percentile threshold for the deltaMu2 statistic.
All genes exceeding this percentile threshold will be selected.}

\item{deltMu1Thresh}{The percentile threshold for the deltaMu1 statistic.
All genes exceeding this percentile threshold will be selected.}

\item{siThresh}{The threshold for the selectivity index statistic (between 0 - 1).
All genes exceeding this threshold will be selected.}
}
\value{
Returns a dataframe containing all genes meeting the prespecified thresholds.
}
\description{
This function allows you to subset genes that are above pre-specified quantiles
and that most closely resemble the distribution of oncogenes.
}
\examples{
topGeneQuants(mmParams)
}
\seealso{
\code{\link{mixModelParams}}
}
\keyword{subsetting}
